Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -4,19 +4,19 @@ int main(){
     char c;
     scanf("%d %d %c", &a, &b, &c);
     if (c=='+'){
-        float sum = a+b;
+        int sum = a+b;
         printf("%.0f", sum);
     } 
     else if (c == '-'){
-        float sub = a-b;
+        int sub = a-b;
         printf("%.0f", sub);
     }
     else if (c == '*'){
-        float mul = a*b;
+        int mul = a*b;
         printf("%.0f", mul);
     }
     else {
-            float d = a / b;
+            int d = a / b;
             if (d>0){
                 printf("%.0f", d);
             }
